Refactors selectors.js -> creates new file (message-utils) to store all the non-state functions for messages in (because they should not be in the selectors.js)
